Update base for Update on "[Inductor Cutlass backend] (2/2) Remove Cutlass EVT epilogue fusion"
* Removes the Cutlass EVT fusion code, but keeps all other Cutlass backend improvements

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx peterbell10 ipiszy yf225 chenyang78 muchulee8 aakhundov ColinPeppler

[ghstack-poisoned]

+++ test/inductor/test_max_autotune.py
@@ -498,8 +498,8 @@ class TestMaxAutotune(TestCase):
         def mm(a, b):
             return a @ b
 
-        a = torch.randn((128, 256), dtype=torch.float32, device="cuda")
-        b = torch.randn((256, 128), dtype=torch.float32, device="cuda")
+        a = torch.randn((128, 256), dtype=torch.float16, device="cuda")
+        b = torch.randn((256, 128), dtype=torch.float16, device="cuda")
 
         with tempfile.TemporaryDirectory() as tmpdir:
             conf_patch = {
